Removed duplicate entries by just merging these classes
